% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_protection.R
\name{example_protection}
\alias{example_protection}
\title{Launch a small graveler example  for protection monitoring dashboard}
\usage{
example_protection()
}
\description{
An illustration of graveler applied to a  humanitarian protection
            monitoring dashboard
}
\examples{
if (interactive()) {

 example_protection()

}
}
