% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_response.R
\name{example_response}
\alias{example_response}
\title{Launch a small graveler example for response dashboard}
\usage{
example_response()
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 example_response()

}
}
