% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodude_example.R
\name{geodude_example}
\alias{geodude_example}
\title{Launch a small graveler example}
\usage{
geodude_example()
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 geodude_example()

}
}
