% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graveler-package.R
\docType{package}
\name{graveler-package}
\alias{graveler}
\alias{graveler-package}
\title{graveler: Builds standardized, modular shiny dashboard frameworks}
\description{
Wrapper for {golem} that automates dashboard file generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edouard-legoupil/graveler}
}

}
\keyword{internal}
