% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graveler_dashboard.R
\name{graveler_dashboard}
\alias{graveler_dashboard}
\title{Golem dashboard backend}
\usage{
graveler_dashboard(path, ...)
}
\arguments{
\item{path}{path to file}

\item{...}{other parameters}
}
\description{
Creates golem directory and header from template info
}
