% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_up.R
\name{level_up}
\alias{level_up}
\title{Graveler Module}
\usage{
level_up(name, pkg = get_graveler_wd(), open = TRUE, dir_create = TRUE)
}
\arguments{
\item{name}{name}

\item{pkg}{package name}

\item{open}{boolean}

\item{dir_create}{boolean}
}
\description{
Creates module of the specified name
}
