% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_example.R
\name{unhcr_example}
\alias{unhcr_example}
\title{Launch a small graveler example}
\usage{
unhcr_example()
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 unhcr_example()

}
}
