% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_protection.R
\name{example_protection}
\alias{example_protection}
\title{Launch a small graveler example  for protection dashboard}
\usage{
example_protection()
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 example_protection()

}
}
