% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_response.R
\name{example_response}
\alias{example_response}
\title{Launch a small graveler example for response tracking dashboard}
\usage{
example_response()
}
\description{
An illustration of graveler applied to a humanitarian response 
            tracking dashboard
}
\examples{
if (interactive()) {

 example_response()

}
}
